+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Security
 
 
+## [1.1.20] - 2019-11-05
+### Added
+- Added the $_is_force_del_and_send flag in AdvancedSendUtils::showAnyMessage().
+- Added the AdvancedSendUtils::forceShowAnyMessage() method.
+
 ## [1.1.19] - 2019-11-05
 ### Changed
 - Logging have been improved.